Improved the clipping method.
Improved the clipping method + added support for entities.
Instead of teleporting the player to the blocks position, the players heights increases by the distance needed to reach the block. This way it is more accurate and doesn't glitch/bug all that often. (This command will only be used for clipping downwards)